requirements.txt
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
idna==3.4
numpy==1.25.2
pycparser==2.21
PyNaCl==1.5.0
requests==2.31.0
urllib3==2.0.4
